Fix 'The command line is too long' errors on Windows
If all arguments are made absolute the system command line easily
get's too long, which will lead to weird build errors later on.

This can be easily worked around by resolving the path names on
the callee side. Anyhow, we now have to make sure that the file
separators are consistent.

Task-number: QTBUG-53618
Change-Id: I1b28cc43dd890f6c9b723af9b70d2b2614651ba2
Reviewed-by: Michael Brüning <michael.bruning@theqtcompany.com>