Don't cancel an IME composition when a modifier key is pressed.
When an asian IME is used to compose a new word, and a modifier key is
pressed (shift, alt, etc), WebEngine would notify Chromium to cancel
the composition, thus clearing the text. But the actual IME window
would still be present, and if another "letter" key was pressed, the
composition would continue with the previous pre-edit text, which leads
to unusual text flickering.

The previous behavior was introduced in 31efe25d14 to fix a Windows
double character input bug.

The current change makes sure to clear the IME composition only in case
the last received QInputMethodEvent pre-edit and commit strings were
empty, which is not the case when pressing a modifier key for instance.

Change-Id: Ic968404c90e1e0eb703fe1c2849990467bedd5e1
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>