Commit the done-so-far IME composition on mouse click.
Before this change, it was possible to select parts of the pre-edit
text in an ongoing IME composition, by click-dragging with the mouse
(which also looked wrong, because the styling of the text did not
change to a selection box to signal selection).

This is inconsistent with how Chrome does it, which commits the
pre-edit string entered so far, when clicking anywhere in the content
area or in the input.

This change makes sure to commit the pre-edit string when a mouse
click is done anywhere on the displayed web page.
The behavior was present on Windows and Linux.

Change-Id: I9cb148c591b5d09fb4dd477ae96c29ca32cc34de
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>