Fix backspace to work when removing last IME char.
Currently when composing a word using an IME, and backspace
is pressed to remove the last character present in the pre-edit
string, instead of the character being removed, nothing happens.
In reality the character is removed, but Chromium isn't notified
of it, and the change will be seen only after another key is
pressed.

Fix consists in notifying Chromium to cancel the IME composition
when both the pre-edit string and the commit string are empty.

There is still an issue with japanese, when trying to
input "aaa", press space, then "b" and you don't see
the "b" until you press one more key.

Change-Id: Idf2ef4888caead26d19eabbbdf4f98fbee601049
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>