Handle IME events which contain both preedit and commit strings.
Currently if a QInputMethodEvent is received, which contains text in
both the pre-edit and commit strings, Chromium would only show the
commit string in the text input (or other html element). The IME though
still knows about the non-empty pre-edit string, which means that
if another key is pressed, Chromium will suddenly show the content of
the previous pre-edit string AND the result of the new key press.

To fix this, WebEngine will now properly set the pre-edit string as the
new composition, after confirming the previous commit string.

Change-Id: If22dd2038aca35a6fe6bb58a521f0a7124c7d468
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>