Always show caret when editing an IME pre-edit string.
Previously while entering a Japanese succession of characters, followed
by transforming a substring into a word suggestion, the caret would
disappear, and pressing the arrow keys would not indicate which part of
the pre-edit string will be transformed next.

This change makes sure the caret is always present, to indicate which
part of the pre-edit string will be replaced by a possible IME
suggestion.

Change-Id: I350310c198bcacf0bcb48217f84b08e95ab8e8ef
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>