Prevent a crash after having downloaded a file
Since the profile owns the QWebEngineDownloadItem then it is likely that
it will still be around when the QWebEngineProfile is deleted. As the
QWebEngineDownloadItem is a child of the QWebEngineProfile then it will
try to delete it when deleting the QWebEngineProfile which means it cannot
trigger a function call into QWebEngineProfilePrivate.

Change-Id: I51077a7857fb49a6708224a9e9942d17de6f6778
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>