Override short-cuts to common editor commands
To ensure Chromium is given a chance to handle editor commands,
we must override these short-cuts.

On OS X we must also perform the action afterwards as these are not
handled internally by the Blink Editor.

The patch solves copy/paste in flash plugins and copy/paste on OS X
when no application short-cuts have been defined. The handling of
short-cut override events is based on how it was handled in Qt WebKit

Task-number: QTBUG-54221
Change-Id: I748671c7bfa5662aae16c6a4b9bbe5e2bce1b907
Reviewed-by: Alexandru Croitor <alexandru.croitor@theqtcompany.com>