Remove handling of DIR_MEDIA_LIBS
DIR_MEDIA_LIBS was where ffmpegsumo was before 5.6 which removed it.
When we don't have the library it used to be in, we return the fallback
path, which by OverridePath will be created if it doesn't exits.

DIR_MEDIA_LIBS is not used anywhere by Chromium anymore, so remove the
code.

Task-number: QTBUG-54258
Change-Id: I1015bd3549c617916c06b633ba3e483ef855ff84
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>