Clear internal selected text when searching
Previously if a selection was made on a web page, and afterwards a find
operation is executed, the selection in the web page would be cleared,
but the call to selectedText() would still return the old selection.

Make sure selectedText() is always cleared, when starting a find
operation, as well as when stopping one.

Change-Id: If78f0fa1dd836a52184015e749ef5a84b9f784cd
Task-number: QTBUG-54071
Reviewed-by: Michael Brüning <michael.bruning@qt.io>