Fix building on OS X 10.9 with 10.10 SDK
Previously building on OS X 10.9 was only allowed with a 10.10.3 SDK.
But there is no supported version of Xcode on 10.9 that would ship with
that SDK, which lead to the fact that there was no way to compile
WebEngine with an officialy provided toolchain.

This patch lowers the requirement of the SDK to 10.10, at the expense
of disabling usage of API that was added in the 10.10.3 SDK release
(Force Touch API).

The required minimum Xcode version is thus bumped to 6.1, and the
documentation is updated accordingly.

Task-number: QTBUG-54486
Change-Id: I025caa336ceac5b8ea76ef451eb5e6b78abfe0c9
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>