Fix issue where Chromium thought it had active focus
Previously, when a WebEngineView's RenderWidgetHostViewQtDelegateQuick
had focus, but not activeFocus, Chromium was told that it had keyboard
focus. That is not correct, because having focus inside the FocusScope
(the WebEngineView), does not mean it has keyboard focus, it just
means the RWHVQDQ will receive active focus, when the WebEngineView
does.

The call path for that erronous check was done when a new page was
loaded, specifically going through
NavigatorImpl::DidNavigate() ->
RenderFrameHostManager::CommitPending() ->
render_frame_host_->GetView()->HasFocus() ->
RenderWidgetHostViewQtDelegateQuick::hasKeyboardFocus(),
which resulted in incorrect display of a blinking caret in an
input, when the view actually wasn't focused.

The fix consists in checking for the activeFocus rather than the
focus property.

Change-Id: I97002be3fc8adfec2228a1b5ad62696f8872d3cc
Reviewed-by: Michael Brüning <michael.bruning@qt.io>