Add persistent backend to channel id service
Channel-ids are only supposed to be memory only when running in off
the record profiles. We have just never initialized the sql-based
backend.

We follow the cookie-settings, because channel-ids are used together
with cookies, have similar implications, and newer Chromium versions
will assert that cookie-store and channel-id store have matching
storage models.

Change-Id: I0a64146f0ed36a8913706bfc3fcadd7404894745
Reviewed-by: Michael Brüning <michael.bruning@qt.io>