Clicking on a select box option, did not work
There was introcuded a new logic which proves that
no mouse / keyboard events are forwarded to
Chromium if the view has no focus, and activeFocusOnPress is set to false.

The selection box get focus when the user click on it,
but the popup window never get (focus is on QuickRootItem)
in this case these mouse events are ignored to forward to Chromium.

Task-number: QTBUG-54795
Change-Id: Id6e81ee39dcde21a6c5c46e302888b9e9478352f
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>