Fix IME window placement relative to QWebEngineView
The position of the IME window is computed using the widget input item
transform. When a regular QWidget gets a focusIn event, the input item
transform is recomputed inside the
QWidgetPrivate::updateWidgetTransform method. This did not happen
for the QWebEngineView, because the focus event is handled internally
and not passed down to QWidget::event.
Fix consists in calling updateWidgetTransform manually whenever the
view receives focus.
The other cases when updateWidgetTransform should be called (namely
resize and move events) are handled properly by delegating to
QWidget::event.

Task-number: QTBUG-55634
Change-Id: Ic93662929e169d860f8ca567f1955da4dc45f9fe
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>