Fix select tag interaction when the web view is inside a modal dialog
Previously when a QWebEngineView was inside a modal QDialog, trying to
click on a select tag option did not properly select the option. It
either focused the new option without closing the popup, or didn't
focus it at all.

Fix consists in making sure the newly created popup QWindow and
RenderWidgetHostViewQtDelegateWidget are marked as children of the
QWebEngineView, so that they are considered part of the current modal
session by the OS, thus allowing user interaction with them.

Because the ownership of the delegate widget should still be retained
by its respective RenderWidgetHostViewQt instance, the QObject parent
of the delegate is unset before the parent is destroyed.

Also to make it work on macOS, the window attribute has to be set
to Qt::Tool instead of Qt::ToolTip.

Change-Id: I56d6f446254a624428a0c661ac3c49eb409c931e
Task-number: QTBUG-54836
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Michael Brüning <michael.bruning@qt.io>