Fix namespace build on macOS
The namespace build was broken because both Qt and Chromium try to
forward declare NSString in certain cpp files. The forward declarations
were incompatible, because the Qt NSString was inside a namespace which
is brought into scope with "using QtNamespace::NSString", whereas the
Chromium one isn't in a namespace, and that leads to ambiguity issues.

Fix is not to use QT_FORWARD_DECLARE_CLASS which prepends the Qt
namespace, but rather do a simple forward declaration.

Change-Id: I8cd5fd0a9a2f62643f6afbd4508df885e13c9ae5
Reviewed-by: Michael Brüning <michael.bruning@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>