Pepper flash search-path for package adobe-flashplugin
Since version 54, the chromium distribution package does not contain
the pepper flash plugin anymore. Because of this, the plugin cannot
be installed using the Ubuntu package "pepperflashplugin-nonfree"
anymore (because the package pulled the plugin from the chromium
distribution).
The plugin is also contained in the package adobe-flashplugin, which
installs the file to /usr/lib/adobe-flashplugin/libpepflashplayer.so.

[ChangeLog] Pepper flash is now also searched in /usr/lib/adobe-flashplugin/

Change-Id: I7c135a2eac1628b0d338bbfed07695c9578458e6
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>