Fix font load error messages inside non-app-bundle tests
Some fonts were failing to load in tests because macOS does not
consider some font paths when an application is built without an
Info.plist file (as is the case for a non-bundle application).
It is possible though to embed a plist file into the executable by
passing the path to the file as a linker argument.

This change generates an Info.plist for each test,  and embeds it into
the final executable, thus fixing the font problems.

Task-number: QTBUG-57354
Change-Id: I4c3c29442b9d308ee6a327645054a67c4b008ef8
Reviewed-by: Jake Petroules <jake.petroules@qt.io>