Prevent accessing d when view closed during DnD operation
When a WebEngineView is closed while a Drag-n-Drop operation
is in progress, the this pointer and its d-pointer become
invalid, so the accesses after drag->exec must be prevented
from being executed. To achieve this, a lambda function is
connected to dragSources &QObject::destroyed signal, which
invalidates a bool on the stack (and cancels the Drag-n-Drop
operation).

Task-number: QTBUG-57713
Change-Id: I9cbb5e6aba99e307d62773bc18f4fec5f79cf703
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>