Replace ASSERT in FaviconManager::setIcon by return
The ASSERT in FaviconManager::setIcon can be triggered by what seems
to be a race condition. Apparently the old request can be gone (so
m_inProgressRequests.contains(id) is false), but m_inProgressRequests
can still be non-empty, because of a new request to the same URL, so
this case is not covered by checking m_inProgressRequests.isEmpty()
before.

Task-number: QTBUG-57900
Change-Id: I2f26c05e5c97e4bd8d1cea03e8005cf61f2b0c98
Reviewed-by: Michael Brüning <michael.bruning@qt.io>