Add https to list of internal schemes
Since https is an internal scheme, it should be in this list of
internal schems.
This also prevents custom custom QWebEngineUrlSchemeHandlers to
register "https". Although it is debatable whether this is necessary,
https should be held to at least the same security standards as http,
which is also in that list.

Task-number: QTBUG-56894
Change-Id: I83d0130541fe3b5831bf29d429bad89a8fc5408c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>