Fix building spellchecker example when Qt is not yet installed
During the build of the spellchecker example, the dictionary
conversion tool is executed to convert some dictionaries. The tool
will fail to find the ICU data file because the file is not present
at the Qt install location at examples build time.

The fix is to allow passing the ICU data directory via an environment
variable. Specifically the QtWebEngine build directory which contains
the ICU data file will be passed when the spellchecker example is
built.

Task-number: QTBUG-58451
Change-Id: I6188f0b7aabe8b2e55cd2e0d59553282058c5035
Reviewed-by: Michal Klocek <michal.klocek@qt.io>