Provisional Fix for Navigation-Breakage on right-click in Flash applet
When one right-clicks into a flash applet, chromium believes that a
context menu was open (even though
PepperRendererHostFactoryQt::CreateResourceHost returns NULL, which
only causes an ASSERT on chromiums side).
Because of this, the navigation is disabled, although no context menu
is displayed. This fix just sends a message to chromium, telling it
that "the context menu" is now "closed", which causes another ASSERT,
but re-enables the navigation.

Task-number: QTBUG-57924
Change-Id: Id6d8150a7d4944162bab81362153f8621d35bb4c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>