Fix QWebEngineView::setFocus to properly set internal QQuickItem focus
The widgets object hierarchy related to focus goes like this:
QWebEngineView's focus proxy is ->
RenderWidgetHostViewQtDelegateWidget, which has an internal
QQuickRootItem defined by QQuickWidget, and the child of the item is ->
RenderWidgetHostViewQuickItem.

Previously when QWebEngineView::setFocus was called, the focus was set
on the RenderWidgetHostViewQtDelegateWidget and the QQuickRootItem,
but not on the RenderWidgetHostViewQuickItem. This caused for e.g.
an active HTML text input not receiving focus.

Make sure the RenderWidgetHostViewQuickItem is marked to have focus
within its root item, so that if the root item receives active focus,
so will RenderWidgetHostViewQuickItem receive it.

Task-number: QTBUG-58515
Change-Id: I175610e3dfebc03733aefe26c16f47096df8ff5b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>