Yet another speculative fix of linkHovered QML tests
When the test moves the mouse cursor out of the link area it positions
the cursor above an iframe. We should not expect linkHovered signal
above an iframe until its load is succeeded. Since we don't have API to
test iframes' loading state, remove the iframe from the test.
Moreover, increase signal timeout values to 12sec as it is used in other
tests too since the 55-based chromium update.

Task-number: QTBUG-55870
Change-Id: Id47494ac41a9d34f8a5bf4804fc5c157868cf923
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>