Fix crashes in removeFromCanvas
After reparenting the view, the compositor data may still be the
same, but the node might be a different one. Check if the list of
scene graph nodes is empty when making the decision on whether to
build a new node tree or not. Also advance the node iterator for
missing render passes.

Task-number: QTBUG-58472
Change-Id: I28dfb40bda89470c7cdaec0aef8a976e307f5b6a
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>