Add QPdfPageNavigation class
The QPdfPageNavigation class encapsulates the logic of navigating
inside a QPdfDocument. It has the notion of a 'current' page and
provides slots to navigate to the previous/next page or jump
directly to a given page.
Additionally it provides properties to indicate whether there is
a previous or next page, relative to the current one.

Change-Id: I053f3c49ab4a70b2610b64d96d2c274c3d0f629b
Reviewed-by: Marc Mutz <marc.mutz@kdab.com>