Ignore synthetic hover move events if the mouse has not actually moved
6f84a09dfb and b2501b0bda in qtdeclarative introduced a synthetic
hover move  event being sent once per frame if the item underneath
the mouse is marked as dirty. This triggers a loop when the web
contents changes reacting to mouse move events as the item needs to
be marked dirty to be rerendered.

Ignore hover move events if the position of the mouse has actually not
changed to work around this.

Task-number: QTBUG-58561
Change-Id: Ic7f3c06e7acf5dfeacc95e347bd026233c957556
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>