Add support for macOS Airplay to Demobrowser in presence of dual GPUs
Previously when Airplay was activated and a new QWebEngineView instance
was created, a black area would be rendered inside the view.
This happened because of incompatible OpenGL contexts. The global Qt
OpenGL context could not share its textures with the newly created
"after-airplay" context, due to the contexts being used by different
GPUs (only possible on a MacBook Pro).

The fix is to add the NSSupportsAutomaticGraphicsSwitching option into
the application plist file, which informs macOS that the application
supports usage of dual GPUs, and thus allows sharing of contexts
across GPUs.

Task-number: QTBUG-54053
Change-Id: Ic37fb75f164922bcca6456054174171fa9ba2bfb
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>