Don't use_external_popup_menu on macOS so that popups actually appear
Using the use_external_popup_menu option forces Chromium to compile
a different code path for displaying popup menus on macOS, which is
not implemented by the Qt overrides, and thus no popups appear.

Thus that option should be disabled, and instead the old code path
should be used, just like in previous versions of Qt.

Change-Id: Ic0f0dca2e097aac2c01e729d7a91425ce3f52674
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>