Pass <input type="file"> result to correct RenderFrameHost
In WebContentsDelegateQt::RunFileChooser, we get a
content::RenderFrameHost*, to which chromium wants us to pass the
files that were selected. We now store that pointer in the
FilePickerController, so that we can pass the files there in
FilePickerController::filesSelectedInChooser.
The attribute content::WebContents *m_contents is no longer needed
then, because that was only used to obtain the main frame to which
the files were sent (this caused problems when the file input was
inside an iframe - we sent the result to the wrong frame in that
case, so the file input object remained blocked).

Task-number: QTBUG-59168
Change-Id: Ic2fdacb225e5a1a9e46ced6a1e4a9ff428f18a62
Reviewed-by: Michael Brüning <michael.bruning@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Michal Klocek <michal.klocek@qt.io>