Implement proper touch handling on macOS
Use the OS provided QNativeGestureEvents for pinching and smart
zooming. This replaces the usage of the Android based gesture
recognizer on macOS.

This also implements multitouch gestures to work with the Qt Quick
implementation of QtWebEngine for macOS, because touch events
are ignored by default on macOS in QQuickItem.

Task-number: QTBUG-58779
Change-Id: I17399e4e89a57557540b2fd0940a445326b682f3
Reviewed-by: Florian Bruhin <qt-project.org@the-compiler.org>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>