Make FocusOnNavigationEnabled setting disabled by default
Before this setting was added in Qt 5.8, the behavior was that
QWebEngineView/WebEngineView got the focus automatically. With the new
setting added it is configurable, but it was left enabled for
compatibility reasons. It would had to be changed at some point anyway:
none of other QML items or widgets grab the focus after creation, so why
would QWebEngineView/WebEngineView do?
This patch also fixes a bug with the text cursor in WebView (the cursor
was blinking as if WebView had the focus when it did not).

[ChangeLog][Important Changes][QtWebEngine][General]
focusOnNavigationEnabled setting which allows controlling whether
a web view will receive focus on a navigation request is now disabled
by default.

Task-number: QTBUG-60152
Task-number: QTBUG-60149
Change-Id: I78dc9bb9ffc70bf06217952acd456b97651c5185
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Michal Klocek <michal.klocek@qt.io>