Fix crash on macOS when closing last tab of a window
Because of weird interaction between deleteLater(),
Qt::WA_DeleteOnClose, and QQuickRenderControlPrivate::windowDestroyed()
which executes all the deferred deletions at a bad time, there were
some use-after-free crashes.

The workaround is not to close the window when the last tab is closed,
but rather create a new tab, like the behavior was on non-macOS
platforms.

Task-number: QTBUG-58982
Change-Id: I00295838c298af382cac78bb22c47d6de60e7755
Reviewed-by: Michal Klocek <michal.klocek@qt.io>