Disable hardware video decoding on macOS
When WebEngine is built with proprietary codecs enabled, this leads
to usage of the macOS video hardware acceleration framework. Because
of missing support of Rectangle OpenGL textures on the Qt WebEngine
side, AVC videos are not displayed correctly.

Disable hardware video decoding on macOS until said support is
implemented.

Task-number: QTBUG-60002
Change-Id: I5faf8df00362d442c5a6f207cc2a9637d663ad00
Reviewed-by: Michael Brüning <michael.bruning@qt.io>