Fix selectedText() to return the value of what findText("") finds
Previously when a find operation was finished by calling findText(""),
the selection buffer was cleared and calling selectedText() would
return an empty string, even though on the UI side a blue rectangle
selection was still visible.

This was due to an incorrect Unselect() call in
WebContentsAdapter::stopFinding().

With the new selection change support in the parent patch, and with the
removed Unselect() call, selectedText() now properly returns the value
of the highlighted blue selection after searching is finished.

Task-number: QTBUG-60673
Task-number: QTBUG-54071
Task-number: QTBUG-53134
Change-Id: I89e0eddb0c14af6d6c06ee878e706be65d3b0831
Reviewed-by: Viktor Engelmann <viktor.engelmann@qt.io>
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>