Fix qwebengine_convert_dict to work with macOS framework build
The ICU data file in a macOS framework build is located inside the
QtWebEngineCore.framework/Resources directory. The dictionary
conversion tool looked for the ICU data file only in the installed
Qt location directory, which is incorrect, thus leading to a conversion
failure.

The fix is to check for the data file in the QtWebEngineCore framework
directory.

Task-number: QTBUG-60707
Change-Id: Ic2325e2b5a5db4e7cbcea3bed92755d5f8b92f95
Reviewed-by: Michal Klocek <michal.klocek@qt.io>