Remove duplicate installation of resource files
The resources files are currently installed twice. Once in src/core.pro
and once in src/core_module.pro.

The core.pro installation steps should be removed, because they do not
cover the macOS framework case, thus ending up with creating resources
both in the framework directory, as well as in the installation root.

Task-number: QTBUG-60707
Change-Id: Id0121c82d7439949474f605c6630ae045125f7ba
Reviewed-by: Michal Klocek <michal.klocek@qt.io>