Fix QMimeData to content::DropData conversion
When files are dropped then it's expected that dataTransfer.files and
dataTransfer.items have the same amount of entries. That means that
whenever QMimeData contains files we must only convert those and ignore
the rest.

But while files do not allow other items before them, text and URLs may
co-exist in content::DropData.

This is consistent with the behavior of Chromium and Firefox and what
web pages expect.

Task-number: QTBUG-59690
Change-Id: Iad56aac3e071dcfa09fa5dfcc788c1c54c6eeae6
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>