Don't disable local storage for offTheRecord profiles
Chromium has local storage enabled in icognito windows, and simply doesn't
persist it to disk. QtWebEngine seems to do the same when local storage is
enabled in an off-the-record profile, so there's no reason to disable it by
default.

This also matches the behavior in Chromium (and Firefox, but not Safari).

[ChangeLog] HTML 5 local storage is now enabled by default (but residing in
memory) in off-the-record profiles.

Change-Id: I0cb7e946575a53471ffed9d3324b3ae4c2e80eee
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>