Fix WebRTC screen sharing on macOS
Previously when a screen sharing request was accepted on macOS, instead
of showing the captured screen, a black rectangle was displayed. This
was due to passing an incorrect screen id to the screen capturer, which
resulted in a failed attempt to find the main active screen.

The fix is to query for all available screen ids from the operating
system, and choose the first id as the id to pass along to the screen
capturer.

Task-number: QTBUG-55165
Change-Id: Id8e648e59755aa2820b05b990adeaa9b58fd26f0
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>