Fix typos in symbol names
The symbol name for delete operators did not
match the actual function definitions, however
linker had no problem with that and simply skipping
the symbol version export. Due to lazy binging
in PLT this also got unnoticed when testing
with test app, where delete was called on exit.

Task-number: QTBUG-60565
Change-Id: I2be2f394a0097f1d06cf9ecda33ae5b22f83a77c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>