Implement geometryChangeRequested signal in the Quick API
Chromium includes the size of the frame in the requested geometry,
but QWindow::setGeometry() and the related properties of QML Window
expect a size without frame. This patch
 - fixes the geometry handling in Demobrowser example.
 - implements a new signal in the QML API which sends both frameless
and extended geometries to the application, because QML doesn't have
capability to determine the size of the frame and handle extended
geometry correctly.

Task-number: QTBUG-51181
Change-Id: I8595593aecca33d7cd46986c7abbad0cc7c18ec2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>